2007-11-01  Hidehisa Akiyama  <akky@users.sourceforge.jp>
	* src/arm.h:
	* src/audio.cc:
	* src/bodysender.cc:
	* src/coach.h:
	* src/field.cpp:
	* src/fullstatesender.cc:
	* src/initsender.cc:
	* src/monitor.h:
	* src/object.cpp:
	* src/object.h:
	* src/pcombuilder.h:
	* src/pcomparser.h:
	* src/player.cc:
	* src/player.h:
	* src/referee.cc:
	* src/serverparam.h:
	* src/types.h:
	* src/vector.h:
	* src/visual.cc:
	* src/visual.h:
	- cleaned up.
	- added a new state flag, POST_COLLIDE.



git-svn-id: file:///home/akiyama/sserver/svn/rcss/trunk/rcssserver@29 1639d2a8-b03b-4162-8b0b-03f2cde17881

+++ src/serverparam.h
@@ -952,7 +952,7 @@ public:
     const double & tackleBackDist() const { return M_tackle_back_dist; }
     const double & tackleWidth() const { return M_tackle_width; }
     const double & tackleExponent() const { return M_tackle_exponent; }
-    unsigned int tackleCycles() const { return (unsigned int)M_tackle_cycles; }
+    int tackleCycles() const { return M_tackle_cycles; }
     const double & tacklePowerRate() const { return M_tackle_power_rate; }
     const double & minTacklePower() const { return M_min_tackle_power; }
     const double & maxTacklePower() const { return M_max_tackle_power; }
